% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KD.R
\name{KD}
\alias{KD}
\title{Kernel (h-mode) depth for functional data}
\usage{
KD(X, X0, u_vec, multi = FALSE)
}
\arguments{
\item{X}{An n by p matrix of in-sample functional data, where each row represent one observed curve. The depth is constructed by this dataset.}

\item{X0}{An n0 by p matrix of in-sample functional data, where each row represent one observed curve. The depth is evaluated at these data points. If X0=NULL, only the in-sample depth values are provided.}

\item{u_vec}{A numerical vector of quantile levels to determine the bandwidths for kernel depth.}

\item{multi}{A logical vector. If TRUE, it is applied for functional data. Otherwise, it is applied for multivariate data.}
}
\value{
A list of kernel depth values and selected bandwidths.
\item{KD}{The kernel depth values of both in-sample and out-of-sample functional data over the quantile levels in u_vec}
\item{h_vec}{The selected bandwidths given the quantile levels in u_vec}
}
\description{
Compute the kernel depth values of in-sample and out-of-sample functional data over different quantile levels
}
\examples{
# example =========

# simulate the regressor curves from Karhunen-Loeve expansion and responses
J=15

# eigenvalues
c=2; a=3.5; dt = c*(1:J)^(-a)
g1 = c*VGAM::zeta(a)
ga = c(g1, sapply(1:(J-1), function(j){g1 - sum(dt[1:j])}))

# eigenfunctions
tt = 70; tGrid = 1:tt/tt; J=15
phi= t(fda::fourier(tGrid, J))

n=101
xi = matrix(rnorm(n*J), n, J)
X = xi \%*\% (phi * sqrt(ga))
n0=50
xi0 = matrix(rnorm(n0*J), n0, J)
X0 = xi0 \%*\% (phi * sqrt(ga))

u_vec = c(0.5, 0.3, 0.05, 0.01,0.001)

resKD = KD(X, X0, u_vec)
}
\references{
Cuevas, Febrero, and Fraiman (2006) On the use of the bootstrap for estimating functions with functional data. Computational Statistics & Data Analysis, 51(2):1063--1074 

Wynne and Nagy (2025) Statistical depth meets machine learning: Kernel mean embeddings and depth in functional data analysis. International Statistical Review, 93(2):317--348
}
\seealso{
\code{\link{RegDepth}}
\code{\link{depthInfer2test}}
\code{\link{depthInferFoFR}}
}
