% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegDepth.R
\name{RegDepth}
\alias{RegDepth}
\title{Regarulized halfspace and projection depths for functional data}
\usage{
RegDepth(X, X0, u_vec, Mproj = 10000, verbose = FALSE)
}
\arguments{
\item{X}{An n by p matrix of in-sample functional data, where each row represent one observed curve. The depth is constructed by this dataset.}

\item{X0}{An n0 by p matrix of in-sample functional data, where each row represent one observed curve. The depth is evaluated at these data points. If X0=NULL, only the in-sample depth values are provided.}

\item{u_vec}{A numerical vector of quantile levels to determine the regularizations.}

\item{Mproj}{An integer. The total number of random projections used to approximate the regularized halfspace and projection depths. For each quantile level u in u_vec, (1-u)*Mproj directions will be used for constructing the depth functions.}

\item{verbose}{A logical vector. If FALSE, only depth and RegPara appear in the result. Otherwise, it also shows shows regularized indices, projections of the functional data, halfspace probabilties, projection outlyingness, generated projection directions, standard deviations and median absolute deviations of the projected data.}
}
\value{
\item{depth}{The regularized halfspace and projection depth values of both in-sample and out-of-sample functional data over the quantile levesl in u_vec by regularizing either standard deviations or median absolute deviations of the projected data}
\item{RegPara}{The selected regularizations by the quantile levels in u_vec.}
\item{whichReg}{The indices of the regularized projections.}
\item{Xproj}{The projected functional data onto the generated random directions.}
\item{HP}{The halfspace probabilities.}
\item{PO}{The projection outlyingness.}
\item{ProjDir}{The randomly generated directions.}
\item{ProjDispersion}{The standard deviations and median absolute deviations of the projected data.}
}
\description{
Compute the regularized halfspace and projection depth values of in-sample and out-of-sample functional data over different quantile levels
}
\examples{
# example =========

# simulate the regressor curves from Karhunen-Loeve expansion and responses
J=15

# eigenvalues
c=2; a=3.5; dt = c*(1:J)^(-a)
g1 = c*VGAM::zeta(a)
ga = c(g1, sapply(1:(J-1), function(j){g1 - sum(dt[1:j])}))

# eigenfunctions
tt = 70; tGrid = 1:tt/tt; J=15
phi= t(fda::fourier(tGrid, J))

n=101
xi = matrix(rnorm(n*J), n, J)
X = xi \%*\% (phi * sqrt(ga))
n0=50
xi0 = matrix(rnorm(n0*J), n0, J)
X0 = xi0 \%*\% (phi * sqrt(ga))

u_vec = c(0.5, 0.3, 0.05, 0.01,0.001)
resRD = RegDepth(X, X0, u_vec)

}
\references{
Yeon, H., Dai, X., and Lopez-Pintado, S. (2025) Regularized halfspace depth for functional data. Journal of the Royal Statistical Society Series B: Statistical Methodology, 87(5):1553--1575

Bočinec, F., Nagy, S., and Yeon, H. (2026+) Projection detph for functional data: Theoretical properties. Under major revision

Yeon, H. (2026+) Effective and flexible depth-based inference for functional parameters. In preparation
}
\seealso{
\code{\link{KD}}
\code{\link{depthInfer2test}}
\code{\link{depthInferFoFR}}
}
